Fix segfault on strcmp and fix process ID
When finding the position of the '.' in the file extension, there would be times that strchr would return null. Throwing that into the strcmp on line 138 would result in a segfault.

Also the process ID has been tweaked. We dont need to do anything with the currentProcessChildCounter interms of returns, we will only deal with exit statuses and a single increment of the sum of the exit status to get the full count.