Getting closer but running grade still shows fail
When testing this out outside of the greader it appears to work fine when not funnedled to an output.txt file. Test1,2,3 all work well even when funneled to an output.txt file, but Test 4,5,6 still have issues with a funneled STDOUT.

The grading still says the PIDs are wrong and im not sure why. Maybe PIDs have to be in order?